moved all interface classes into folder
